[FLINK-12963] [state-processor] Simplify SavepointMetadata hierarchy
From the current state of its usage, the previous hierarchy
(ModifiableSavepointMetada -> SavepointMetadata) is redundant. This
commit simplifies this to just be a single SavepointMetadata class.
